refactor: extract framebuffer ops to petroleum crate
Move scroll_buffer_pixels and clear_buffer_pixels functions from fullerene-kernel to the petroleum crate for reuse. Update imports and calls in graphics.rs to use the shared implementations, reducing code duplication across modules.
